bee-packable crate refactoring
