+++ bee-common/bee-common-derive/Cargo.toml
@@ -18,7 +18,7 @@ name = "tests"
 path = "tests/lib.rs"
 
 [package.metadata.cargo-udeps.ignore]
-development = ["bee-common"]
+development = ["bee-packable"]
 
 [dependencies]
 proc-macro-error = "1.0"
@@ -27,7 +27,7 @@ quote = "1.0"
 syn = { version = "1.0", features = ["full", "extra-traits"] }
 
 [dev-dependencies]
-bee-common = { path = "../bee-common" }
+bee-packable = { path = "../bee-packable" }
 
 rustversion = { version = "1.0" }
 trybuild = { version = "1.0", features = ["diff"] }
