+++ bee-common/bee-common-derive/tests/packable/fail/invalid_tag_type_enum.rs
@@ -3,7 +3,7 @@
 
 #![allow(unused_imports)]
 
-use bee_common::packable::Packable;
+use bee_packable::Packable;
 
 use core::convert::Infallible;
 
@@ -14,7 +14,7 @@ pub enum OptI32 {
     #[packable(tag = 0)]
     None,
     #[packable(tag = 1)]
-    Some(i32)
+    Some(i32),
 }
 
 fn main() {}
