+++ bee-common/bee-common-derive/tests/packable/fail/packable_is_structural.stderr
@@ -9,7 +9,7 @@ error[E0599]: no function or associated item named `pack` found for struct `NonP
    |
    = help: items from traits can only be used if the trait is implemented and in scope
    = note: the following trait defines an item `pack`, perhaps you need to implement it:
-           candidate #1: `bee_common::packable::Packable`
+           candidate #1: `bee_packable::Packable`
    = note: this error originates in a derive macro (in Nightly builds, run with -Z macro-backtrace for more info)
 
 error[E0599]: no function or associated item named `unpack` found for struct `NonPackable` in the current scope
@@ -23,7 +23,7 @@ error[E0599]: no function or associated item named `unpack` found for struct `No
    |
    = help: items from traits can only be used if the trait is implemented and in scope
    = note: the following trait defines an item `unpack`, perhaps you need to implement it:
-           candidate #1: `bee_common::packable::Packable`
+           candidate #1: `bee_packable::Packable`
    = note: this error originates in a derive macro (in Nightly builds, run with -Z macro-backtrace for more info)
 
 error[E0599]: no function or associated item named `packed_len` found for struct `NonPackable` in the current scope
@@ -37,5 +37,5 @@ error[E0599]: no function or associated item named `packed_len` found for struct
    |
    = help: items from traits can only be used if the trait is implemented and in scope
    = note: the following trait defines an item `packed_len`, perhaps you need to implement it:
-           candidate #1: `bee_common::packable::Packable`
+           candidate #1: `bee_packable::Packable`
    = note: this error originates in a derive macro (in Nightly builds, run with -Z macro-backtrace for more info)
