+++ bee-common/bee-common-derive/tests/packable/pass/generic_newtype.rs
@@ -3,7 +3,7 @@
 
 #![allow(unused_imports)]
 
-use bee_common::packable::Packable;
+use bee_packable::Packable;
 
 #[derive(Packable)]
 #[packable(error = T::Error)]
