+++ bee-common/bee-common-derive/tests/packable/pass/mixed_enum.rs
@@ -3,7 +3,7 @@
 
 #![allow(unused_imports)]
 
-use bee_common::packable::{UnknownTagError, Packable};
+use bee_packable::{error::UnknownTagError, Packable};
 
 #[derive(Packable)]
 #[packable(tag_type = u8)]
@@ -12,7 +12,7 @@ pub enum Foo {
     #[packable(tag = 0)]
     Bar(u32),
     #[packable(tag = 1)]
-    Baz{ x: i32, y: i32 }
+    Baz { x: i32, y: i32 },
 }
 
 fn main() {}
