+++ bee-common/bee-common-derive/tests/packable/pass/named_enum.rs
@@ -3,7 +3,7 @@
 
 #![allow(unused_imports)]
 
-use bee_common::packable::{UnknownTagError, Packable};
+use bee_packable::{error::UnknownTagError, Packable};
 
 #[derive(Packable)]
 #[packable(tag_type = u8)]
@@ -12,7 +12,7 @@ pub enum OptPoint {
     #[packable(tag = 0)]
     None,
     #[packable(tag = 1)]
-    Some{ x: i32, y: i32 }
+    Some { x: i32, y: i32 },
 }
 
 fn main() {}
