+++ bee-common/bee-common-derive/tests/packable/pass/unnamed_enum.rs
@@ -3,7 +3,7 @@
 
 #![allow(unused_imports)]
 
-use bee_common::packable::{Packable, UnknownTagError};
+use bee_packable::{error::UnknownTagError, Packable};
 
 #[derive(Packable)]
 #[packable(tag_type = u8)]
@@ -12,7 +12,7 @@ pub enum OptI32 {
     #[packable(tag = 0)]
     None,
     #[packable(tag = 1)]
-    Some(i32)
+    Some(i32),
 }
 
 fn main() {}
