+++ bee-common/bee-common-derive/tests/packable/pass/unnamed_struct.rs
@@ -3,7 +3,7 @@
 
 #![allow(unused_imports)]
 
-use bee_common::packable::Packable;
+use bee_packable::Packable;
 
 use core::convert::Infallible;
