--- bee-common/bee-common/src/lib.rs
@@ -1,11 +1,0 @@
-// Copyright 2021 IOTA Stiftung
-// SPDX-License-Identifier: Apache-2.0
-
-//! A crate that provides common functionalities shared across multiple crates within the Bee framework, and for
-//! applications built on-top.
-
-#![no_std]
-#![warn(missing_docs)]
-
-/// A module that provides a `Packable` trait to serialize and deserialize types.
-pub mod packable;
