+++ bee-common/bee-packable/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
-name = "bee-common"
-version = "1.0.0-alpha"
+name = "bee-packable"
+version = "0.1.0"
 authors = ["IOTA Stiftung"]
 edition = "2018"
 description = "Common utilities used across the bee framework"
